Exercise 4.21-4.22

+++ part4/bloglist/controllers/blogs.js
@@ -17,6 +17,9 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  if(!request.token){
+    return response.status(401).json({ error: 'token missing' })
+  }
   const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
@@ -38,14 +41,37 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  if(!request.token){
+    return response.status(401).json({ error: 'token missing' })
+  }
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  const blog = await Blog.findById(request.params.id)
+  const createdUserId = blog.user.toString()
+  if(user.id.toString() === createdUserId){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else{
+    return response.status(403).json({ error: 'not the correct user' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
+  if(!request.token){
+    return response.status(401).json({ error: 'token missing' })
+  }
   const body = request.body
 
-  const user = await User.findById(body.userId)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
 
   const blog = {
     title: body.title,
@@ -55,7 +81,7 @@ blogsRouter.put('/:id', async (request, response) => {
     user: user._id
   }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new : true })
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new : true , runValidators : true })
   response.json(updatedBlog)
 })
